# docs/swagger.yaml
basePath: /api/v1/
definitions:
...
  domain.EmailInput:
    properties:
      email:
        type: string
    required:
    - email
    type: object
...
  domain.GetAllUsersCategoryResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/domain.UserList'
        type: array
      page-info:
        $ref: '#/definitions/domain.PaginationPage'
    type: object
...
  domain.PaginationPage:
    properties:
      count:
        type: integer
      page:
        type: integer
      pages:
        type: integer
    type: object
...
  domain.UserList:
    properties:
      age:
        type: integer
      city:
        type: string
      email:
        type: string
      gender:
        type: string
      id:
        type: integer
      is_blocked:
        type: boolean
      name:
        type: string
      phone_number:
        type: string
      registered_at:
        type: number
    type: object
...
  v1.AdminSignUpInput:
    properties:
      confirm_password:
        type: string
      email:
        type: string
      name:
        type: string
      password:
        maxLength: 64
        minLength: 8
        type: string
      phone_number:
        type: string
    required:
    - confirm_password
    - email
    - name
    - password
    - phone_number
    type: object
...
  v1.NotificationForUser:
    properties:
      link:
        type: string
      text:
        type: string
      title:
        type: string
      user_ids:
        items:
          type: integer
        type: array
    required:
    - text
    - title
    - user_ids
    type: object
...
  v1.ResetPasswordInput:
    properties:
      new_password:
        maxLength: 64
        minLength: 8
        type: string
      new_password_confirm:
        type: string
      token:
        type: string
    required:
    - new_password
    - new_password_confirm
    - token
    type: object
  v1.TokenAuth:
    properties:
      token:
        type: string
    type: object
  v1.UserSignUpInput:
    properties:
      age:
        type: integer
      city:
        type: integer
      confirm_password:
        type: string
      email:
        type: string
      gender:
        maximum: 2
        minimum: 1
        type: integer
      name:
        type: string
      password:
        maxLength: 64
        minLength: 8
        type: string
      phone_number:
        type: string
    required:
    - age
    - city
    - confirm_password
    - email
    - gender
    - name
    - password
    - phone_number
    type: object
  v1.idResponse:
    properties:
      id: {}
    type: object
  v1.okResponse:
    properties:
      message:
        type: string
    type: object
  v1.refreshInput:
    properties:
      token:
        type: string
    required:
    - token
    type: object
  v1.response:
    properties:
      detail:
        type: string
    type: object
  v1.signInInput:
    properties:
      password:
        type: string
      phone_number:
        type: string
    required:
    - password
    - phone_number
    type: object
  v1.tokenResponse:
    properties:
      access:
        type: string
      refresh:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: fiber@swagger.io
    name: API Support
  description: API Server for  A Hundred to Five
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: A Hundred to Five
  version: "2.0"
paths:
  /admin/sign-up:
    post:
      consumes:
      - application/json
      description: create Admin
      parameters:
      - description: admin sign-up
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/v1.AdminSignUpInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/v1.idResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.response'
      tags:
      - auth
  /auth/block-user/{id}:
    put:
      consumes:
      - application/json
      description: block users
      parameters:
      - description: user id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.okResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.response'
      security:
      - User_Auth: []
      tags:
      - auth
  /auth/download/users:
    get:
      description: download members to excel
      operationId: downloadUsers
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.response'
      tags:
      - auth
  /auth/refresh:
    post:
      description: refresh token
      parameters:
      - description: refresh info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/v1.refreshInput'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.tokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.response'
      tags:
      - auth
  /auth/reset-password:
    post:
      consumes:
      - application/json
      description: reset password
      parameters:
      - description: 'reset password '
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/v1.ResetPasswordInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/v1.okResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.response'
      tags:
      - auth
  /auth/sign-in:
    post:
      consumes:
      - application/json
      description: user sign in
      parameters:
      - description: sign in info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/v1.signInInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.tokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.response'
      tags:
      - auth
  /auth/sign-up:
    post:
      consumes:
      - application/json
      description: create member account
      parameters:
      - description: user sign-up
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/v1.UserSignUpInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/v1.idResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.response'
      tags:
      - auth
  /auth/unblock-user/{id}:
    put:
      consumes:
      - application/json
      description: unblock users
      parameters:
      - description: user id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.okResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.response'
      security:
      - User_Auth: []
      tags:
      - auth
  /auth/user/me:
    get:
      consumes:
      - application/json
      description: get users info
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.UserList'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.response'
      security:
      - User_Auth: []
      tags:
      - auth
  /auth/users:
    get:
      consumes:
      - application/json
      description: gets all users
      operationId: getAllUsers
      parameters:
      - in: query
        minimum: 0
        name: limit
        type: integer
      - in: query
        minimum: 0
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.GetAllUsersCategoryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.response'
      tags:
      - auth
  /auth/verify:
    post:
      consumes:
      - application/json
      description: reset password
      parameters:
      - description: reset password via Email
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/domain.EmailInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/v1.TokenAuth'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.response'
      tags:
      - auth
...
  /notification/user:
    post:
      consumes:
      - application/json
      description: create notification for specific user
      operationId: createNotyForSpecificUser
      parameters:
      - description: notification
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/v1.NotificationForUser'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/v1.idResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.response'
      security:
      - User_Auth: []
      tags:
      - notification
...
securityDefinitions:
  User_Auth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"